Moved data.semparse to semparse
