--- allennlp/data/semparse/knowledge_graphs/__init__.py
@@ -1,1 +1,0 @@
-from allennlp.data.semparse.knowledge_graphs.table_question_knowledge_graph import TableQuestionKnowledgeGraph
