--- allennlp/data/semparse/type_declarations/__init__.py
@@ -1,1 +1,0 @@
-from allennlp.data.semparse.type_declarations.grammar_state import GrammarState
