--- allennlp/data/semparse/worlds/__init__.py
@@ -1,2 +1,0 @@
-from allennlp.data.semparse.worlds.nlvr_world import NlvrWorld
-from allennlp.data.semparse.worlds.wikitables_world import WikiTablesWorld
