+++ allennlp/semparse/__init__.py
@@ -3,6 +3,6 @@ This module contains classes related to handling data for semantic parsing. That
 declarations, and world representations for all the domains for which we want to build semantic parsers.
 """
 
-from allennlp.data.semparse.knowledge_graphs.knowledge_graph import KnowledgeGraph
-from allennlp.data.semparse.worlds.world import ParsingError, World
-from allennlp.data.semparse.action_space_walker import ActionSpaceWalker
+from allennlp.semparse.knowledge_graphs.knowledge_graph import KnowledgeGraph
+from allennlp.semparse.worlds.world import ParsingError, World
+from allennlp.semparse.action_space_walker import ActionSpaceWalker
