+++ allennlp/semparse/knowledge_graphs/knowledge_graph.py
