+++ allennlp/semparse/type_declarations/__init__.py
@@ -1,0 +1,1 @@
+from allennlp.semparse.type_declarations.grammar_state import GrammarState
