+++ allennlp/semparse/type_declarations/type_declaration.py
