+++ allennlp/semparse/worlds/__init__.py
@@ -1,0 +1,2 @@
+from allennlp.semparse.worlds.nlvr_world import NlvrWorld
+from allennlp.semparse.worlds.wikitables_world import WikiTablesWorld
