+++ tests/models/encoder_decoders/wikitables_semantic_parser_test.py
@@ -8,12 +8,12 @@ from torch.autograd import Variable
 from allennlp.common import Params
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.common.testing import ModelTestCase
-from allennlp.data.semparse.type_declarations import GrammarState
-from allennlp.data.semparse.type_declarations.type_declaration import START_SYMBOL
 from allennlp.models import Model, WikiTablesSemanticParser
 from allennlp.models.encoder_decoders.wikitables_semantic_parser import WikiTablesDecoderState
 from allennlp.models.encoder_decoders.wikitables_semantic_parser import WikiTablesDecoderStep
 from allennlp.modules import SimilarityFunction
+from allennlp.semparse.type_declarations import GrammarState
+from allennlp.semparse.type_declarations.type_declaration import START_SYMBOL
 
 class WikiTablesSemanticParserTest(ModelTestCase):
     def setUp(self):
